refactor: standardize file/module naming conventions across repository
